+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Support for complex object fields in settings configuration.
+- Support for matching a command pattern `!command pattern <name> [pattern]`.
 
 [Unreleased]: https://github.com/udoprog/OxidizeBot/compare/0.4.4...master